# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Servidores",
        "description": "API Servidores",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/servidores-efetivos": {
            "get": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Lista todos os servidores efetivos",
                "description": "Retorna uma lista paginada de todos os servidores efetivos.",
                "operationId": "ab7898ecc4c601caa12e55678460c603",
                "responses": {
                    "200": {
                        "description": "Lista de servidores efetivos retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ServidorEfetivoResource"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Nenhum servidor efetivo encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Cria um novo servidor efetivo",
                "description": "Cria um novo registro de servidor efetivo.",
                "operationId": "f31386d346e3c66518e168868c398e0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreServidorEfetivoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Servidor efetivo criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServidorEfetivoResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados de entrada inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servidores-efetivos/{id}": {
            "get": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Busca um servidor efetivo por ID",
                "description": "Retorna os detalhes de um servidor efetivo específico com base no seu ID.",
                "operationId": "1992d3377aa6eebe37a5aa5e90a65a77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do servidor efetivo a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Servidor efetivo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServidorEfetivoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servidor efetivo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Atualiza um servidor efetivo existente",
                "description": "Atualiza os dados de um servidor efetivo específico com base no seu ID.",
                "operationId": "d34e13244dc492ec6d3449adc4f289ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do servidor efetivo a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateServidorEfetivoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Servidor efetivo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServidorEfetivoResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados de entrada inválidos"
                    },
                    "404": {
                        "description": "Servidor efetivo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Servidores Efetivos"
                ],
                "summary": "Exclui um servidor efetivo por ID",
                "description": "Remove um servidor efetivo específico com base no seu ID.",
                "operationId": "e31ba3a27537cb0234ebaf2354f16f63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do servidor efetivo a ser excluído",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Servidor efetivo excluído com sucesso"
                    },
                    "404": {
                        "description": "Servidor efetivo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servidores-efetivos/{id}/enderecos": {
            "post": {
                "tags": [
                    "Servidores Efetivos",
                    "Endereços"
                ],
                "summary": "Adiciona um endereço a um servidor efetivo existente",
                "description": "Adiciona um novo endereço à servidor efetivo especificada.",
                "operationId": "6000693913ed74ee04f50f671fb77a93",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do servidor efetivo para adicionar o endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddEnderecoToServidorEfetivoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereço adicionado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServidorEfetivoResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Servidor Efetivo não encontrado"
                    },
                    "400": {
                        "description": "Requisição inválida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servidores-efetivos/unidade/{unidadeId}": {
            "get": {
                "tags": [
                    "Busca"
                ],
                "summary": "Lista os servidores efetivos por ID da unidade",
                "description": "Retorna uma lista de servidores efetivos pertencentes a uma unidade específica.",
                "operationId": "9f1e66d694689f462545f71543810520",
                "parameters": [
                    {
                        "name": "unidadeId",
                        "in": "path",
                        "description": "ID da unidade para buscar os servidores efetivos",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de servidores efetivos da unidade retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ConsultaServidorLotadoPorUnidadeResource"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Nenhum servidor efetivo encontrado para esta unidade"
                    },
                    "404": {
                        "description": "Unidade não encontrada"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/servidores-efetivos/endereco-funcional": {
            "get": {
                "tags": [
                    "Busca"
                ],
                "summary": "Consulta endereço funcional de servidores efetivos por nome (parcial ou completo)",
                "description": "Retorna uma lista de endereços de unidades de servidores efetivos cujo nome corresponde ao parâmetro de busca.",
                "operationId": "65a33a48ed98599239178b7d2c3567eb",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome (ou parte do nome) do servidor efetivo a ser buscado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de endereços funcionais de servidores efetivos encontrados por nome retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ConsultaEnderecoFuncionalServidorEfetivoPorNomeResource"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "Nenhum servidor efetivo encontrado com este nome"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddEnderecoToServidorEfetivoRequest": {
                "required": [
                    "end_tipo_logradouro",
                    "end_logradouro",
                    "end_numero",
                    "end_bairro",
                    "cid_id"
                ],
                "properties": {
                    "end_tipo_logradouro": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "end_logradouro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "end_numero": {
                        "type": "integer"
                    },
                    "end_bairro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "cid_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "StoreServidorEfetivoRequest": {
                "required": [
                    "pes_nome",
                    "pes_data_nascimento",
                    "pes_sexo",
                    "se_matricula"
                ],
                "properties": {
                    "pes_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "pes_data_nascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "pes_sexo": {
                        "type": "string",
                        "maxLength": 1
                    },
                    "pes_mae": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "pes_pai": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "se_matricula": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "endereco_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UpdateServidorEfetivoRequest": {
                "properties": {
                    "pes_nome": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "pes_data_nascimento": {
                        "type": "string",
                        "format": "date",
                        "nullable": true
                    },
                    "pes_sexo": {
                        "type": "string",
                        "maxLength": 1,
                        "nullable": true
                    },
                    "pes_mae": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "pes_pai": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "se_matricula": {
                        "type": "string",
                        "maxLength": 50,
                        "nullable": true
                    },
                    "endereco_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ConsultaEnderecoFuncionalServidorEfetivoPorNomeResource": {
                "properties": {
                    "pes_id": {
                        "type": "integer"
                    },
                    "pes_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "unid_id": {
                        "type": "integer"
                    },
                    "unid_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "unid_sigla": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "end_id": {
                        "type": "integer"
                    },
                    "end_tipo_logradouro": {
                        "type": "string",
                        "maxLength": 50
                    },
                    "end_logradouro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "end_numero": {
                        "type": "integer"
                    },
                    "end_bairro": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "cid_id": {
                        "type": "integer"
                    },
                    "cid_nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "cid_uf": {
                        "type": "string",
                        "maxLength": 2
                    }
                },
                "type": "object"
            },
            "ConsultaServidorLotadoPorUnidadeResource": {
                "properties": {
                    "pes_id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "idade": {
                        "type": "integer"
                    },
                    "unidade": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "fotografia": {
                        "type": "string",
                        "format": "uri"
                    }
                },
                "type": "object"
            },
...
            "ServidorEfetivoResource": {
                "properties": {
                    "pes_id": {
                        "type": "integer"
                    },
                    "pes_nome": {
                        "type": "string"
                    },
                    "pes_data_nascimento": {
                        "type": "string",
                        "format": "date"
                    },
                    "pes_sexo": {
                        "type": "string",
                        "maxLength": 1
                    },
                    "pes_mae": {
                        "type": "string",
                        "nullable": true
                    },
                    "pes_pai": {
                        "type": "string",
                        "nullable": true
                    },
                    "se_matricula": {
                        "type": "string"
                    },
                    "enderecos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "end_id": {
                                    "type": "integer"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}